requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
base58==2.1.1
    # via -r requirements.in
betterproto[compiler]==2.0.0b4
    # via -r requirements.in
black==22.1.0
    # via betterproto
certifi==2021.10.8
    # via requests
cffi==1.15.0
    # via pynacl
charset-normalizer==2.0.9
    # via requests
click==8.0.3
    # via black
grpclib==0.4.2
    # via betterproto
h2==4.1.0
    # via grpclib
hpack==4.0.0
    # via h2
hyperframe==6.0.1
    # via h2
idna==3.3
    # via requests
jinja2==2.11.3
    # via betterproto
markupsafe==2.0.1
    # via jinja2
multidict==6.0.2
    # via grpclib
mypy-extensions==0.4.3
    # via black
pathspec==0.9.0
    # via black
platformdirs==2.5.0
    # via black
pycparser==2.21
    # via cffi
pynacl==1.5.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via betterproto
requests==2.26.0
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
tomli==2.0.1
    # via black
urllib3==1.26.7
    # via requests
